backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyjwt
filelock
